refact add xml annotation to IRepository to improve expected behavior
this is a wip, as I'm not sure what is the best way to go  regardind that matter (how to handle exceptions thrown by a class that implements an interface, on the class that consumes the interface).
this will probably be changed later on, but I'm satisfied for now with this approach.